Fixes on impact ids; use common class for idu and gidd sources
